requirements.txt
mysql-connector-python==8.2.0
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
colorlog==6.8.0
